[PGS] 몫 구하기 / level 0
https://school.programmers.co.kr/learn/courses/30/lessons/120805?language=cpp
