+++ 3st/Hyeonseo/programmer/숫자 비교하기.cpp
@@ -1,0 +1,14 @@
+#include <string>
+#include <vector>
+
+using namespace std;
+
+int solution(int num1, int num2) {
+    int answer = -1;
+    if(num1 == num2) {
+        answer = 1;
+    }
+    return answer;
+}
+//level 0
+
